Totally rewrote henshin. Now uses Base concept.
Removed the old way where the code was quite disjointed and unified it
under one class (to rule them all). It controls everything and it is
possible (and encouraged) to create subclasses of it for specifiec
problems.

It is not recommended to use any of this as the whole dsl doesn't feel
quite right yet. But it is better than previously.